Feat: Add ability to kill the mob
When the mob dies after its health reaches 0 when hit by a Projectile, its RigidBody3D physics govern its movement and it'll then be deleted after its DeleteTimer runs out.
Many collision layers and some nodes are adjusted, a HitBox and a HurtBox help decouple the code between the projectile and the Mob.