Feat: Add Health component
Entities like Enemy can have a Health component attached to it.
The Health scene contains values: health and max_health
Signals are defined in this component to make it easy to react to value changes be it for a UI updating based on these values or a node which has animations to play depending on if the health is increased, decreased...
The take_damage() function in Enemy is now replaced by this component and it connects to the Health signals instead to play the hurt animation or die.