Feat: Add game over with UI
An autoload Events is used as an events bus. It contains game_started and game_over signals.
game_started is emitted so GameStats autoload resets the score to 0 when the sscene is loaded and the ScoreUI updates according to it.
game_over is emitted when the player dies. PauseMenu is disabled, GameOverUI pops up, PlayAgainButton is focused and if clicked it reloads the current scene (Level) and GameStats are then reset again because game_started is emitted by level when it's ready. That's the important loop with these events to know about.
Also, EnemySpawner stops spawning enemies who disable their physics_process themselves as well.